Small re-organization
